# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: KSA Attendance System API
  version: 0.2.0
  summary: September 2026 implementation baseline between the production frontend and backend.
  description: |
    This contract captures approved API capabilities and stable domain outcomes for
    implementation. Endpoint names may change through a documented superseding decision, but the
    authorization, state, privacy, idempotency, and source-of-truth guarantees must remain.

    The permanent QR points to a stable frontend route such as `/check-in`; it is not an
    API credential. The frontend obtains an explicit browser location and submits it to
    the backend, which makes the final attendance decision using server time.

    The production browser uses the same public origin for the UI and `/api/*`. Cookie-authenticated
    state changes also require the implementation's CSRF/origin control. Abuse-sensitive operations
    may return the documented `RATE_LIMITED` response and safe `Retry-After` under DEC-036.
servers:
  - url: /api/v1
    description: Same-origin API
...
security:
  - sessionCookie: []
paths:
...
  /me/manual-verifications:
    get:
      tags: [Manual verification]
      summary: List the current participant's own manual-verification cases without photo access
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReviewStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Own case status, including pending expiry and final attendance where available
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ManualVerificationCase"
                  nextCursor:
                    type: [string, "null"]
        "403":
          $ref: "#/components/responses/AuthorizationError"
...
  /manual-verification-requests:
    post:
      tags: [Manual verification]
      summary: Explicitly request review after a denied/unavailable/no-fix or clearly-remote location outcome
      description: Never creates a case for inactive/closed sessions, duplicate attendance, or an unrecognized device. Repeated requests return the existing pending case.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attemptId, reason]
              properties:
                attemptId:
                  type: string
                  format: uuid
                reason:
                  type: string
                  minLength: 3
                  maxLength: 500
      responses:
        "200":
          description: Existing pending case returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualVerificationCaseEnvelope"
        "201":
          description: Pending case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualVerificationCaseEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
  /manual-verifications:
    get:
      tags: [Manual verification]
      summary: List manual-verification cases visible to the current operator
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReviewStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Manual cases with minimum list-view participant data
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ManualVerificationCase"
                  nextCursor:
                    type: [string, "null"]
        "403":
          $ref: "#/components/responses/AuthorizationError"
  /manual-verifications/{caseId}:
    get:
      tags: [Manual verification]
      summary: Get a review case including authorized short-lived/proxied photo access
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualVerificationCaseEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /manual-verifications/{caseId}/approve:
    post:
      tags: [Manual verification]
      summary: Approve a pending case and idempotently create manual attendance
      parameters:
        - $ref: "#/components/parameters/CaseId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Approved case and attendance, including already-final result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualVerificationCaseEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /manual-verifications/{caseId}/reject:
    post:
      tags: [Manual verification]
      summary: Reject a pending manual-verification case
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Rejected case or already-final result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualVerificationCaseEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /manual-verifications/emergency-attendance:
    post:
      tags: [Manual verification]
      summary: Create controlled emergency manual attendance during an eligible session
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId, sessionId, reason]
              properties:
                participantId:
                  type: string
                  format: uuid
                sessionId:
                  type: string
                  format: uuid
                reason:
                  type: string
                  minLength: 3
                  maxLength: 1000
      responses:
        "201":
          description: Manual attendance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecordEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
...
  /admin/attendance/corrections:
    post:
      tags: [Administration]
      summary: Apply a reasoned historical attendance correction with before/after audit
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceCorrectionRequest"
      responses:
        "200":
          description: Corrected authoritative attendance state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceCorrectionEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ksa_session
      description: Secure, HttpOnly, host-only, SameSite=Lax server session; raw value is never stored. Idle expiry is 30 days and absolute expiry is 90 days.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AuthorizationError:
      description: Current actor is not allowed to perform this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundError:
      description: Resource is absent or intentionally undisclosed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ConflictError:
      description: Current domain state conflicts with the requested command
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimitedError:
      description: Too many requests; retry according to safe response guidance
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Role:
      type: string
      enum: [PARTICIPANT, COURSE_REP, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]
...
    ParticipantSummary:
      type: object
      required: [id, fullName, serialNumber, status, roles]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        serialNumber:
          type: string
          pattern: "^KSA-[0-9]{2}$"
        status:
          $ref: "#/components/schemas/UserStatus"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    ActorSummary:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
...
    AttendanceMethod:
      type: string
      enum: [QR, MANUAL]
    AttendanceStatus:
      type: string
      enum: [PRESENT, ABSENT, NOT_APPLICABLE]
    AttendanceRecord:
      type: object
      required: [id, sessionId, participant, status, method, checkedInAt]
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        participant:
          $ref: "#/components/schemas/ParticipantSummary"
        status:
          type: string
          enum: [PRESENT]
        method:
          $ref: "#/components/schemas/AttendanceMethod"
        checkedInAt:
          type: string
          format: date-time
        approvedBy:
          oneOf:
            - $ref: "#/components/schemas/ActorSummary"
            - type: "null"
        locationStatus:
          type: string
          enum: [VERIFIED, MANUAL, NOT_RETAINED]
...
    ReviewStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, EXPIRED, SUPERSEDED]
    ManualVerificationCase:
      type: object
      required:
        [
          id,
          participant,
          sessionId,
          creationSource,
          reasonCode,
          status,
          requestedAt,
          expiresAt,
          version,
        ]
      properties:
        id:
          type: string
          format: uuid
        participant:
          $ref: "#/components/schemas/ParticipantSummary"
        sessionId:
          type: string
          format: uuid
        sourceAttemptId:
          type: [string, "null"]
          format: uuid
        creationSource:
          type: string
          enum: [AUTOMATIC_UNCERTAIN, PARTICIPANT_REQUEST, OPERATOR_EMERGENCY]
        reasonCode:
          type: string
          enum:
            [
              LOCATION_UNCERTAIN,
              LOCATION_PERMISSION_DENIED,
              LOCATION_TIMEOUT,
              LOCATION_UNAVAILABLE,
              LOCATION_UNSUPPORTED,
              CLEARLY_REMOTE,
              EMERGENCY,
            ]
        participantReason:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/ReviewStatus"
        requestedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          description: Session close plus 15 minutes
        reviewedAt:
          type: [string, "null"]
          format: date-time
        reviewedBy:
          type: [string, "null"]
          format: uuid
        decisionReason:
          type: [string, "null"]
        photoAccessUrl:
          type: [string, "null"]
          description: Authorized presigned URL lasting only a few minutes; detail response only
        attendanceRecord:
          oneOf:
            - $ref: "#/components/schemas/AttendanceRecord"
            - type: "null"
        version:
          type: integer
          minimum: 1
    ReviewDecisionRequest:
      type: object
      required: [expectedVersion, reason]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 3
          maxLength: 1000
...
    AttendanceCorrectionRequest:
      type: object
      required: [sessionId, participantId, targetStatus, reason]
      properties:
        sessionId:
          type: string
          format: uuid
        participantId:
          type: string
          format: uuid
        targetStatus:
          type: string
          enum: [PRESENT, ABSENT, NOT_APPLICABLE]
        checkedInAt:
          type: [string, "null"]
          format: date-time
        reason:
          type: string
          minLength: 3
          maxLength: 1000
    AttendanceCorrection:
      type: object
      required:
        [
          sessionId,
          participantId,
          effectiveStatus,
          reason,
          correctedAt,
          correctedBy,
          auditEventId,
        ]
      properties:
        sessionId:
          type: string
          format: uuid
        participantId:
          type: string
          format: uuid
        effectiveStatus:
          $ref: "#/components/schemas/AttendanceStatus"
        record:
          oneOf:
            - $ref: "#/components/schemas/AttendanceRecord"
            - type: "null"
        reason:
          type: string
        correctedAt:
          type: string
          format: date-time
        correctedBy:
          type: string
          format: uuid
        auditEventId:
          type: string
          format: uuid
          description: Append-oriented event containing the authoritative before/after history
...
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - STALE_STATE
            - DUPLICATE_EMAIL
            - DUPLICATE_PHONE
            - DUPLICATE_SERIAL
            - ROSTER_ENTRY_NOT_ELIGIBLE
            - ROSTER_CLAIM_CONFLICT
            - LAST_ADMIN_REQUIRED
            - RECENT_REAUTHENTICATION_REQUIRED
            - SELF_APPROVAL_FORBIDDEN
            - SESSION_ALREADY_OPEN
            - SESSION_NOT_OPEN
            - SESSION_CLOSED
            - SESSION_EXTENSION_LIMIT
            - DEVICE_CHANGE_REQUIRED
            - LOCATION_UNCERTAIN
            - LOCATION_PERMISSION_DENIED
            - LOCATION_TIMEOUT
            - LOCATION_UNAVAILABLE
            - LOCATION_UNSUPPORTED
            - CLEARLY_REMOTE
            - MANUAL_VERIFICATION_NOT_ALLOWED
            - MANUAL_VERIFICATION_EXPIRED
            - PHOTO_TOO_LARGE
            - PHOTO_UNSUPPORTED
            - RATE_LIMITED
            - SERVICE_TEMPORARILY_UNAVAILABLE
        message:
          type: string
        correlationId:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    AttendanceRecordEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AttendanceRecord"
    AttendanceCorrectionEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AttendanceCorrection"
    ManualVerificationCaseEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ManualVerificationCase"
...